backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.4
mysqlclient>=2.2
selenium>=4.21
sentence-transformers>=3.0
chromadb>=0.5
requests>=2.32
python-dotenv>=1.0
gunicorn>=22.0
